refactor: implement similar methods on abstract class
