+++ app/modules/authors/repository.py
@@ -1,29 +1,19 @@
 from typing import List, Optional
-from sqlalchemy import select
+
 from fastapi import HTTPException
+from sqlalchemy import select
 from sqlalchemy.orm import Session
+
 from app.interfaces.repository import Repository
 
 from . import schemas
-from .models import AuthorModel
 from .entities import Author
+from .models import AuthorModel
 
 
 class AuthorRepository(Repository):
     def __init__(self, session: Session):
-        self.session = session
-
-    def find(self, id: int) -> Optional[Author]:
-        author = self.session.scalars(
-            select(AuthorModel).where(AuthorModel.id == id)
-        ).first()
-        if author is None:
-            return None
-        return author
-
-    def list(self) -> list[Author]:
-        author_models = self.session.scalars(select(AuthorModel)).all()
-        return [author.to_entity() for author in author_models]
+        super().__init__(session)
 
     def create(self, author: schemas.AuthorCreate):
         author_model = AuthorModel(
@@ -34,13 +24,3 @@ class AuthorRepository(Repository):
         self.session.refresh(author_model)
         self.session.flush()
         return author_model.to_entity()
-
-    def delete(self, author_id: int) -> bool:
-        author = self.session.scalars(
-            select(AuthorModel).where(AuthorModel.id == author_id)
-        ).first()
-        if author is None:
-            return False
-        self.session.delete(author)
-        self.session.commit()
-        return True
